Project now uses cmake for the build-process. On unix systems call make or better "make config" to compile the project. The latter command is for configuring slam6d and its components. Make will create a directory .build which will contain all of cmake's generated files. For configuring the project without ccmake do: "cd build && cmake -i ../"
Directory structure has been modified for easier inclusion in ros:
  "include"   contains all headers and .icc files of slam6d
  "src"       contains all .cc and other assorted source code of slam6d
  "3rdparty"  contains all external projects such as hogman/newmat/glui etc.
  "windows"   contains files needed for compilation on windows machines
  "lib"       contains the generated shared libraries
  "obj"       contains the generated static libraries
  "bin"       contains the generated executables

Source code is organized using subdirectories of "include" and "src". Therefore you must use i.e.:
  #include "slam6d/globals.icc" 
instead of:
  #include "globals.icc"
to succesfully compile your program



git-svn-id: https://svn.code.sf.net/p/slam6d/code/trunk@273 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/grid/line.cc
@@ -1,4 +1,4 @@
-#include "line.h"
+#include "grid/line.h"
 #include <cmath>
 
 /**
@@ -20,6 +20,6 @@ line::line(gridPoint* start, gridPoint* end)
  */
 double line::getLength()
 {
-    return sqrt(pow(start->getX() - end->getX(), 2) +
-		pow(start->getZ() - end->getZ(), 2));
+    return sqrt(pow((double)start->getX() - end->getX(), 2) +
+ 		pow((double)start->getZ() - end->getZ(), 2));
 }
